+++ src/main/java/lotto/collection/Lotto.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.collection;
 
 import lotto.util.LottoValidator;
