+++ src/main/java/lotto/collection/LottoTickets.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.collection;
 
 import java.util.ArrayList;
 import java.util.List;
