+++ src/main/java/lotto/domain/user/UserRepository.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.domain.user;
 
 import java.util.ArrayList;
 import java.util.List;
